requirements.txt
keras==3.7.0
matplotlib==3.9.3
numpy==2.0.2
pandas==2.2.3
pillow==11.0.0
streamlit==1.40.2
tensorflow==2.18.0
